test: mirror src/ layout in tests and drop _coverage suffix
Move every test file into a package tree under tests/unit/ that mirrors src/ (server, agents, agents/art, agents/agentic_search[/strategies], utils, orchestrator, tools/art), each with an __init__.py. Rename the new coverage tests to test_<module>.py and merge the four that overlapped an existing file (validators, model_factory, monitored_tool, skill_loader) into a single file per module. No test behavior change: 842 passed, coverage 98%.
