+++ tests/unit/agents/agentic_search/strategies/test_forced_tool.py
