+++ tests/unit/agents/agentic_search/strategies/test_template_fill.py
