+++ tests/unit/agents/agentic_search/test_template_schema.py
