+++ tests/unit/agents/art/test_specialized_agents_errors.py
