+++ tests/unit/agents/test_context_management.py
