+++ tests/unit/orchestrator/test_registry.py
