+++ tests/unit/server/test_ag_ui_app.py
