+++ tests/unit/server/test_ag_ui_event_processor.py
