+++ tests/unit/server/test_ag_ui_event_strategy.py
