+++ tests/unit/server/test_auth_middleware.py
