+++ tests/unit/server/test_cli.py
