+++ tests/unit/server/test_config.py
