+++ tests/unit/server/test_error_classification.py
