+++ tests/unit/server/test_error_recovery.py
