+++ tests/unit/server/test_exceptions.py
