+++ tests/unit/server/test_logging_config.py
