+++ tests/unit/server/test_response_formats.py
