+++ tests/unit/server/test_validators.py
@@ -4,12 +4,10 @@
 # this file be licensed under the Apache-2.0 license or a
 # compatible open source license.
 
-"""Coverage tests for server.validators.ValidatedRunAgentInput.
+"""Unit tests for request input validation — tool-result size bound (issue #138).
 
-Field/model validators are exercised through full model construction where
-possible; the message-shape branches (missing role/content, non-serializable
-tool result) are exercised by invoking the raw field-validator function with
-lightweight stand-in message objects.
+Verifies that ``ValidatedRunAgentInput`` caps client-returned tool results at the API
+boundary so oversized continuation input can't bloat the model context window.
 """
 
 from __future__ import annotations
@@ -23,6 +21,65 @@ from pydantic import ValidationError
 from server.validators import _MAX_TOOL_RESULT_BYTES, ValidatedRunAgentInput
 
 pytestmark = pytest.mark.unit
+
+
+def _input_with_messages(messages: list) -> ValidatedRunAgentInput:
+    return ValidatedRunAgentInput(
+        thread_id="t1",
+        run_id="r1",
+        messages=messages,
+        state={},
+        tools=[],
+        context=[],
+        forwarded_props={},
+    )
+
+
+class TestToolResultSizeBound:
+    """The client-returned tool-result cap (issue #138)."""
+
+    def test_rejects_oversized_tool_result(self) -> None:
+        """A tool result over the cap is rejected at the API boundary."""
+        huge = "x" * (_MAX_TOOL_RESULT_BYTES + 1)
+        msg = ToolMessage(id="m1", role="tool", content=huge, toolCallId="tc1")
+        with pytest.raises(ValidationError, match="tool result is too large"):
+            _input_with_messages([msg])
+
+    def test_accepts_tool_result_within_bound(self) -> None:
+        """A normal-sized tool result passes."""
+        msg = ToolMessage(id="m1", role="tool", content="ok", toolCallId="tc1")
+        vi = _input_with_messages([msg])
+        assert len(vi.messages) == 1
+
+    def test_accepts_tool_result_at_bound(self) -> None:
+        """A tool result exactly at the cap is allowed (boundary is inclusive)."""
+        exact = "x" * _MAX_TOOL_RESULT_BYTES
+        msg = ToolMessage(id="m1", role="tool", content=exact, toolCallId="tc1")
+        vi = _input_with_messages([msg])
+        assert len(vi.messages) == 1
+
+    def test_non_tool_message_not_bounded(self) -> None:
+        """The cap applies only to tool-role messages, not ordinary user content."""
+        big_user = "x" * (_MAX_TOOL_RESULT_BYTES + 1)
+        msg = UserMessage(id="m1", role="user", content=big_user)
+        vi = _input_with_messages([msg])
+        assert len(vi.messages) == 1
+
+
+# --- additional coverage tests (merged) ---
+# SPDX-License-Identifier: Apache-2.0
+#
+# The OpenSearch Contributors require contributions made to
+# this file be licensed under the Apache-2.0 license or a
+# compatible open source license.
+
+"""Coverage tests for server.validators.ValidatedRunAgentInput.
+
+Field/model validators are exercised through full model construction where
+possible; the message-shape branches (missing role/content, non-serializable
+tool result) are exercised by invoking the raw field-validator function with
+lightweight stand-in message objects.
+"""
 
 
 # Raw (unbound) message field-validator function for direct branch testing.
